flask_novel/requirements.txt
Babel==2.12.1
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
flask-babel==3.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
greenlet==3.0.3
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==2.3.7
zipp==3.16.2
